userDebugLogOn - Exposed setting option through setting UI
I'm not super happy with how this is implemented; QT was being really buggy with binding checkboxes to state.

I ended up making a copy of settings to avoid mutating the original object, and modifying the copy to persist
local UI state, and only updating the original if the form is submitted.